Refactor : User, Auth 분리
